improved local project dir setup

+++ src/EnvironVersionControl.php
@@ -10,16 +10,38 @@ namespace Phoenix;
 class EnvironVersionControl extends AbstractDeployer
 {
 
+    /**
+     * @var string
+     */
     public $environ;
 
+    /**
+     * @var string
+     */
     protected $logElement = 'h3';
 
+    /**
+     * @var GithubClient
+     */
     private $github;
 
+    /**
+     * @var TerminalClient
+     */
     private $terminal;
 
+    /**
+     * @var WHM
+     */
     private $whm;
 
+    /**
+     * EnvironVersionControl constructor.
+     * @param TerminalClient $terminal
+     * @param GithubClient $github
+     * @param WHM|null $whm
+     * @param string $environ
+     */
     function __construct(TerminalClient $terminal, GithubClient $github, WHM $whm = null, $environ = 'live')
     {
         $this->environ = $environ;
@@ -30,6 +52,10 @@ class EnvironVersionControl extends AbstractDeployer
         parent::__construct();
     }
 
+    /**
+     * @return bool|null
+     * @throws \Github\Exception\MissingArgumentException
+     */
     function create()
     {
         $this->mainStr();
@@ -43,11 +69,10 @@ class EnvironVersionControl extends AbstractDeployer
         $upstream_repository = $this->github->repo()->get($args['repo']['name']);
         if (!$upstream_repository)
             return $this->logError("Upstream repository not found.");
-        $clonedRepository = $this->terminal->git()->check($args['repo']['dir']);
         if ($this->environ != 'local') {
-            $sshKey = $this->whm->genkey($args['key']['name'], $args['key_passphrase'], 2048, $args['cPanel_account']['user']);
+            $sshKey = $this->whm->genkey($args['key']['name'], $args['key']['passphrase'], 2048, $args['cPanel_account']['user']);
             if (!empty($sshKey)) {
-                $authorisedKey = $this->whm->authkey($args['key']['name']);
+                $authorisedKey = $this->whm->authkey($args['key']['name'], 'authorize', $args['cPanel_account']['user']);
                 $addedSSHConfig = $this->terminal->ssh_config()->create($args['key']['name'], 'github.com', $args['key']['name'], 'git');
             }
 
@@ -57,21 +82,32 @@ class EnvironVersionControl extends AbstractDeployer
                 'remote_name' => "origin"
             ]);
 
-            if (!$clonedRepository) {
-                $clonedRepository = $this->whm->version_control('clone',
-                    $args['repo']['dir'],
-                    $args['repo']['downstream_name'],
-                    $sourceRepository
-                );
-                if (!$clonedRepository)
-                    $clonedRepository = $this->whm->version_control('get', $args['repo']['dir']);
-            }
+            $clonedRepository = $this->whm->version_control('clone',
+                $args['repo']['dir'],
+                $args['repo']['downstream_name'],
+                $sourceRepository,
+                $args['cPanel_account']['user']
+            );
+            if (!$clonedRepository)
+                $clonedRepository = $this->whm->version_control('get', $args['repo']['dir'], $args['cPanel_account']['user']);
+
         } else {
-            if (!$clonedRepository) {
+            $localDirSetup = $this->terminal->localProjectDirSetup();
+            $localDirSetup->setProjectArgs($args['project']);
+
+            $webDirArgs = [
+                'dir' => $args['repo']['dir'],
+                'owner' => $args['repo']['owner'],
+                'group' => $args['repo']['group'],
+                'purpose' => 'git'
+            ];
+
+            if ($localDirSetup->create($webDirArgs)) {
                 $clonedRepository = $this->terminal->git()->clone([
                     //'url' => str_replace('git@github.com', 'github', $upstream_repository['ssh_url']),
                     'url' => $upstream_repository['ssh_url'],
-                    'worktree_path' => $args['repo']['dir']
+                    'worktree_path' => $args['repo']['worktree'],
+                    'repo_path' => $args['repo']['dir']
                 ]);
             }
         }
@@ -80,9 +116,12 @@ class EnvironVersionControl extends AbstractDeployer
         if (!empty($clonedRepository) && $this->terminal->git()->waitForUnlock($args['repo']['dir'])) {
             if ($args['repo']['worktree'] != $args['repo']['dir'])
                 $createdDotGit = $this->terminal->dotGitFile()->create($args['repo']['worktree'], $args['repo']['dir']);
-            $purgedGit = $this->terminal->git()->purge($args['repo']['dir']);
+            else
+                $createdDotGit = true;
+            if ($this->environ != 'local')
+                $purgedGit = $this->terminal->git()->purge($args['repo']['dir']);
             $createdGitignore = $this->terminal->gitignore()->create($args['repo']['worktree']);
-            if ($clonedRepository)
+            if ($this->environ != 'local')
                 $resetGit = $this->terminal->gitBranch()->reset(['worktree' => $args['repo']['worktree'], 'branch' => 'master']);
         }
 
@@ -101,15 +140,18 @@ class EnvironVersionControl extends AbstractDeployer
         }
 
         $this->terminal->exec('git config --global user.name "James Jones"; git config --global user.email "james.jones@phoenixweb.com.au"');
-        $success = (!empty($sshKey)
-            && !empty($authorisedKey)
-            && !empty($addedSSHConfig)
-            && !empty($uploadedDeployKey)
+        $success = (
+            ($this->environ == 'local' || (
+                    !empty($sshKey)
+                    && !empty($authorisedKey)
+                    && !empty($addedSSHConfig)
+                    && !empty($uploadedDeployKey)
+                    && !empty($purgedGit)
+                    && !empty($resetGit)
+                ))
             && !empty($clonedRepository)
             && !empty($createdDotGit)
             && !empty($createdGitignore)
-            && !empty($purgedGit)
-            && !empty($resetGit)
             && ($this->environ != 'staging' || (!empty($webhook) && !empty($webhook_config)))
         ) ? true : false;
 
@@ -118,6 +160,9 @@ class EnvironVersionControl extends AbstractDeployer
 
     }
 
+    /**
+     * @return bool|null
+     */
     function delete()
     {
         $this->mainStr();
@@ -129,26 +174,39 @@ class EnvironVersionControl extends AbstractDeployer
             return $this->logError("Couldn't get args");
 
 
+        if ($this->environ != 'local') {
+            $downstream_repository = $this->whm->version_control('delete', $args['repo']['dir'], '', '', $args['cPanel_account']['user']);
+            $sshKey = $this->whm->delkey($args['key']['name'], $args['cPanel_account']['user']);
+            $deployKey = $this->github->deploy_key()->remove($args['repo']['name'], $args['key']['title']);
+            $sshConfig = $this->terminal->ssh_config()->delete('github_' . $args['repo']['name']);
+        }
         $gitignore = $this->terminal->gitignore()->delete($args['repo']['worktree']);
-        $downstream_repository = $this->whm->version_control('delete', $args['repo']['dir'], '', '', $args['cPanel_account']['user']);
         $deleted_git_folder = $this->terminal->git()->delete($args['repo']['dir']);
-        $sshKey = $this->whm->delkey($args['key']['name'], $args['cPanel_account']['user']);
-        $sshConfig = $this->terminal->ssh_config()->delete('github_' . $args['repo']['name']);
-        $deployKey = $this->github->deploy_key()->remove($args['repo']['name'], $args['key']['title']);
+
 
         if ($this->environ == 'staging') { //webhook
             $webhook = $this->github->webhook()->remove($args['repo']['name'], $args['webhook']['url']);
             $webhook_config = $this->terminal->githubWebhookEndpointConfig()->delete($args['webhook']['endpoint_config_dir']);
         }
-        $dotGit = $this->terminal->ssh->delete($args['repo']['worktree'] . '/.git');
 
-        $success = (!empty($gitignore)
-            && !empty($downstream_repository)
-            && !empty($sshKey)
-            && !empty($sshConfig)
-            && !empty($deployKey)
+        if ($args['repo']['worktree'] != $args['repo']['dir'])
+            $dotGit = $this->terminal->dotGitFile()->delete($args['repo']['worktree']);
+        else
+            $dotGit = true;
+
+        $success = (
+            ($this->environ == 'local' || (
+                    !empty($downstream_repository)
+                    && !empty($sshKey)
+                    && !empty($deployKey)
+                    && !empty($sshConfig)
+                ))
+            && !empty($gitignore)
             && !empty($deleted_git_folder)
-            && ((!empty($webhook) && !empty($webhook_config)) || $this->environ != 'staging')
+            && ($this->environ != 'staging' || (
+                    !empty($webhook)
+                    && !empty($webhook_config))
+            )
             && !empty($dotGit)
         ) ? true : false;
 
@@ -156,25 +214,34 @@ class EnvironVersionControl extends AbstractDeployer
 
     }
 
+    /**
+     * @return bool
+     */
     function validate()
     {
         return true;
     }
 
+    /**
+     * @return bool
+     */
     protected function getArgs()
     {
         $environ = $this->environ;
-        $root = $this->terminal->api()->root;
+        $root = $this->terminal->root;
         if (empty($root))
             return $this->logError(sprintf("Couldn't get %s environment root directory.", $environ));
-
 
         $args['repo']['name'] = ph_d()->config->version_control->repo_name ?? '';
         if (empty($args['repo']['name']))
             return $this->logError("Repository name is missing from config.");
         $args['repo']['dir'] = ph_d()->get_environ_dir($environ, 'git');
         $args['repo']['downstream_name'] = $args['repo']['name'] . '_website';
-        $args['repo']['worktree'] = ph_d()->get_environ_dir($environ, 'web');
+        $args['repo']['worktree'] = ph_d()->get_environ_dir($environ, 'worktree');
+
+        $args['repo']['owner'] = ph_d()->config->environ->$environ->directory->web->owner ?? '';
+        $args['repo']['group'] = ph_d()->config->environ->$environ->directory->web->group ?? '';
+
 
         if ($environ != 'local') {
             $args['cPanel_account'] = ph_d()->find_environ_cpanel($environ);
@@ -188,11 +255,21 @@ class EnvironVersionControl extends AbstractDeployer
             $args['webhook']['url'] = 'https://' . $args['cPanel_account']['domain'] . '/github-webhook.php?github=yes';
             $args['webhook']['endpoint_config_dir'] = ph_d()->get_environ_dir($environ, 'github_webhook_endpoint_config') . '/' . $args['repo']['name'] . '.json';
             $args['webhook']['secret'] = ph_d()->config->version_control->github->webhook->secret ?? '';
+        } else {
+            $args['project'] = [
+                'dir' => ph_d()->get_environ_dir('local', 'project') ?? '',
+                'owner' => ph_d()->config->environ->$environ->directory->project->owner ?? '',
+                'group' => ph_d()->config->environ->$environ->directory->project->group ?? '',
+            ];
         }
         return $args;
     }
 
     //protected function mainStr(array $args = [])
+
+    /**
+     * @return string
+     */
     protected function mainStr()
     {
         $action = $this->getCaller();
